Cambiar información de producto según el tipo de producto

+++ SmartTrade/SmartTrade/Logica/Entities/Deporte.cs
@@ -9,13 +9,13 @@ namespace SmartTrade.Entities
     {
         public Deporte() { }
         
-        public Deporte(string nombre, string huella, string imagen, string modelo3d, string desc, int puntos, Vendedor vend, Categoria cat, int stock, int precio, string tipo)
+        public Deporte(string nombre, string huella, string imagen, string modelo3d, string desc, int puntos, Vendedor vend, Categoria cat, int stock, double precio, string tipo)
             :base(nombre, huella, imagen, modelo3d, desc, puntos, vend, cat, stock, precio)
         {
             this.Tipo = tipo;
         }
 
-        public Producto CrearProducto(string nombre, string huella, string imagen, string modelo3d, string desc, int puntos, Vendedor vend, Categoria cat, int stock, int precio, string tipo)
+        public Producto CrearProducto(string nombre, string huella, string imagen, string modelo3d, string desc, int puntos, Vendedor vend, Categoria cat, int stock, double precio, string tipo)
         {
             return new Deporte(nombre, huella, imagen, modelo3d, desc, puntos, vend, cat, stock, precio, tipo);
         }
