+++ SmartTrade/SmartTrade/Logica/Entities/Tecnologia.cs
@@ -9,7 +9,7 @@ namespace SmartTrade.Entities
     {
         public Tecnologia() { }
         
-        public Tecnologia(string nombre, string huella, string imagen, string modelo3d, string desc, int puntos, Vendedor vend, Categoria cat, int stock, int precio, string dispositivo, string marca, string modelo)
+        public Tecnologia(string nombre, string huella, string imagen, string modelo3d, string desc, int puntos, Vendedor vend, Categoria cat, int stock, double precio, string dispositivo, string marca, string modelo)
             :base(nombre, huella, imagen, modelo3d, desc, puntos, vend, cat, stock, precio)
         {
             this.Dispositivo = dispositivo;
@@ -17,7 +17,7 @@ namespace SmartTrade.Entities
             this.Modelo = modelo;
         }
 
-        public Producto CrearProducto(string nombre, string huella, string imagen, string modelo3d, string desc, int puntos, Vendedor vend, Categoria cat, int stock, int precio, string dispositivo, string marca, string modelo)
+        public Producto CrearProducto(string nombre, string huella, string imagen, string modelo3d, string desc, int puntos, Vendedor vend, Categoria cat, int stock, double precio, string dispositivo, string marca, string modelo)
         {
             return new Tecnologia(nombre, huella, imagen, modelo3d, desc, puntos, vend, cat, stock, precio, dispositivo, marca, modelo);
         }
